understanding-and-applying-text-embeddings/requirements.txt
python-dotenv
google-cloud-aiplatform
scikit-learn
mplcursors
pandas
tqdm
pyarrow
scann==1.4.0